other(ts): convert flow JS in 'entity-editor/common' to TypeScript
